refactor(runtime): split orchestrator/tests — extract sanitize + misc
Sixth slice in the stack. Pull the remaining named clusters out of
`tests/mod.rs` into two more sibling files:

`tests/sanitize.rs` (9 tests):
- `sanitize_history_empty_is_noop`
- `sanitize_history_valid_alternation_is_noop`
- `sanitize_history_trailing_user_inserts_assistant`
- `sanitize_history_trailing_multimodal_user_inserts_assistant`
- `sanitize_history_orphaned_tool_calls_get_synthetic_results`
- `sanitize_history_fully_orphaned_tool_calls_all_results_inserted`
- `sanitize_history_combined_orphaned_tools_and_trailing_user`
- `sanitize_history_orphaned_tool_result_dropped`
- `sanitize_history_tool_result_after_matched_calls_dropped`

`tests/misc.rs` (8 tests):
- `value_to_params_map_converts_object`
- `value_to_params_map_empty_object`
- `value_to_params_map_non_object_returns_empty`
- `run_turn_max_iterations_returns_error`
- `turn_identity_reaches_tool_handler` (+ `IdentityCaptureTool`)
- `run_turn_marks_had_errors_when_tool_fails` (+ `FailingTool`)
- `run_turn_clears_had_errors_when_no_tool_fails`
- `run_turn_with_tools_marks_had_errors_when_extension_fails`

Also lifts `ollama_thinking` from the now-empty "Thinking persistence
tests" section up into the helpers block alongside `ollama_answer`
(it's used only by sibling modules now).

cargo test -p assistant-runtime --lib reports 220 passed (same as
before); per-submodule breakdown:
- orchestrator::tests          10 (basic history threading)
- orchestrator::tests::bus_worker 12
- orchestrator::tests::end_turn 6
- orchestrator::tests::misc      8
- orchestrator::tests::multimodal 15
- orchestrator::tests::sanitize  9
- orchestrator::tests::streaming 5
- orchestrator::tests::subagent 10
                               ─── 75 total
