+++ crates/runtime/src/orchestrator/tests/mod.rs
@@ -4,10 +4,7 @@ use std::sync::atomic::{AtomicUsize, Ordering};
 
 use assistant_core::tool::{ToolHandler, ToolOutput};
 use assistant_core::types::conversation::{ExecutionContext, Interface, TurnIdentity};
-use assistant_core::{
-    ChatHistoryMessage, ChatRole, ContentBlock, LlmProvider, MessageBus, OrgId, SpaceId,
-    ToolCallItem, UserId, types::agent::AssistantConfig,
-};
+use assistant_core::{LlmProvider, MessageBus, types::agent::AssistantConfig};
 use assistant_llm_provider::ollama::client::{LlmClient, LlmClientConfig};
 use assistant_storage::{StorageLayer, registry::SkillRegistry};
 use assistant_tool_executor::ToolExecutor;
@@ -28,6 +25,16 @@ fn ollama_answer(text: &str) -> Value {
     json!({
         "model": "test",
         "message": { "role": "assistant", "content": text },
+        "done": true
+    })
+}
+
+/// Ollama response with empty content but non-empty thinking — triggers
+/// `LlmResponse::Thinking` in the LLM client parser.
+fn ollama_thinking(thought: &str) -> Value {
+    json!({
+        "model": "test",
+        "message": { "role": "assistant", "content": "", "thinking": thought },
         "done": true
     })
 }
@@ -177,7 +184,9 @@ impl ToolHandler for MockExtTool {
 
 mod bus_worker;
 mod end_turn;
+mod misc;
 mod multimodal;
+mod sanitize;
 mod streaming;
 mod subagent;
 
@@ -633,599 +642,5 @@ async fn three_tool_calls_handled_in_single_iteration() {
         reqs.len(),
         2,
         "three tool calls must be handled in ONE iteration"
-    );
-}
-
-// ── sanitize_history tests ────────────────────────────────────────────────
-
-#[test]
-fn sanitize_history_empty_is_noop() {
-    let mut history = vec![];
-    crate::history::sanitize_history(&mut history);
-    assert!(history.is_empty());
-}
-
-#[test]
-fn sanitize_history_valid_alternation_is_noop() {
-    let mut history = vec![
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            content: "hello".into(),
-        },
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            content: "hi".into(),
-        },
-    ];
-    crate::history::sanitize_history(&mut history);
-    assert_eq!(history.len(), 2, "valid alternation should not be modified");
-}
-
-#[test]
-fn sanitize_history_trailing_user_inserts_assistant() {
-    let mut history = vec![ChatHistoryMessage::Text {
-        role: ChatRole::User,
-        content: "orphaned".into(),
-    }];
-    crate::history::sanitize_history(&mut history);
-    assert_eq!(
-        history.len(),
-        2,
-        "should insert a synthetic assistant message"
-    );
-    match &history[1] {
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            content,
-        } => {
-            assert!(
-                content.contains("error"),
-                "synthetic message should mention error"
-            );
-        }
-        other => panic!("expected Text(Assistant), got {:?}", other),
-    }
-}
-
-#[test]
-fn sanitize_history_trailing_multimodal_user_inserts_assistant() {
-    let mut history = vec![ChatHistoryMessage::MultimodalUser {
-        content: vec![ContentBlock::Text("image msg".into())],
-    }];
-    crate::history::sanitize_history(&mut history);
-    assert_eq!(history.len(), 2);
-    assert!(matches!(
-        &history[1],
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            ..
-        }
-    ));
-}
-
-#[test]
-fn sanitize_history_orphaned_tool_calls_get_synthetic_results() {
-    let mut history = vec![
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            content: "do stuff".into(),
-        },
-        ChatHistoryMessage::AssistantToolCalls(vec![
-            ToolCallItem {
-                name: "tool-a".into(),
-                params: serde_json::json!({}),
-                id: None,
-            },
-            ToolCallItem {
-                name: "tool-b".into(),
-                params: serde_json::json!({}),
-                id: None,
-            },
-        ]),
-        // Only one ToolResult — tool-b is missing.
-        ChatHistoryMessage::ToolResult {
-            name: "tool-a".into(),
-            content: "ok".into(),
-        },
-    ];
-    crate::history::sanitize_history(&mut history);
-    // Should have: User, AssistantToolCalls, ToolResult(a), ToolResult(b-synthetic)
-    assert_eq!(history.len(), 4, "missing tool result should be inserted");
-    match &history[3] {
-        ChatHistoryMessage::ToolResult { name, content } => {
-            assert_eq!(name, "tool-b");
-            assert!(
-                content.contains("lost") || content.contains("crash") || content.contains("error"),
-                "synthetic result should indicate failure: {content}"
-            );
-        }
-        other => panic!("expected ToolResult, got {:?}", other),
-    }
-}
-
-#[test]
-fn sanitize_history_fully_orphaned_tool_calls_all_results_inserted() {
-    let mut history = vec![
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            content: "run tools".into(),
-        },
-        ChatHistoryMessage::AssistantToolCalls(vec![
-            ToolCallItem {
-                name: "alpha".into(),
-                params: serde_json::json!({}),
-                id: None,
-            },
-            ToolCallItem {
-                name: "beta".into(),
-                params: serde_json::json!({}),
-                id: None,
-            },
-        ]),
-        // No ToolResult at all — process crashed right after persisting tool calls.
-    ];
-    crate::history::sanitize_history(&mut history);
-    // Should have: User, AssistantToolCalls, ToolResult(alpha), ToolResult(beta)
-    assert_eq!(
-        history.len(),
-        4,
-        "both missing tool results should be inserted"
-    );
-    assert!(matches!(&history[2], ChatHistoryMessage::ToolResult { name, .. } if name == "alpha"));
-    assert!(matches!(&history[3], ChatHistoryMessage::ToolResult { name, .. } if name == "beta"));
-}
-
-#[test]
-fn sanitize_history_combined_orphaned_tools_and_trailing_user() {
-    // Simulates: process crashed during tool execution on turn 1,
-    // then on turn 2 the user message was persisted but LLM failed.
-    let mut history = vec![
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            content: "turn 1".into(),
-        },
-        ChatHistoryMessage::AssistantToolCalls(vec![ToolCallItem {
-            name: "my-tool".into(),
-            params: serde_json::json!({}),
-            id: None,
-        }]),
-        // Missing ToolResult, then orphaned user from turn 2:
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            content: "turn 2".into(),
-        },
-    ];
-    crate::history::sanitize_history(&mut history);
-    // Should have: User, AssistantToolCalls, ToolResult(synthetic), User, Assistant(synthetic)
-    assert_eq!(history.len(), 5);
-    assert!(
-        matches!(&history[2], ChatHistoryMessage::ToolResult { name, .. } if name == "my-tool")
-    );
-    assert!(matches!(
-        &history[4],
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            ..
-        }
-    ));
-}
-
-#[test]
-fn sanitize_history_orphaned_tool_result_dropped() {
-    // Simulates: a system-injected tool result (e.g. skill-learner) appears
-    // at the start of history with no preceding AssistantToolCalls.
-    let mut history = vec![
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            content: "hello".into(),
-        },
-        ChatHistoryMessage::ToolResult {
-            name: "skill-learner".into(),
-            content: "Auto-created skill 'foo'".into(),
-        },
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            content: "hi".into(),
-        },
-    ];
-    crate::history::sanitize_history(&mut history);
-    // The orphaned ToolResult should be dropped
-    assert_eq!(history.len(), 2);
-    assert!(matches!(
-        &history[0],
-        ChatHistoryMessage::Text {
-            role: ChatRole::User,
-            ..
-        }
-    ));
-    assert!(matches!(
-        &history[1],
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            ..
-        }
-    ));
-}
-
-#[test]
-fn sanitize_history_tool_result_after_matched_calls_dropped() {
-    // Extra ToolResult beyond what the tool calls declared should be dropped.
-    let mut history = vec![
-        ChatHistoryMessage::AssistantToolCalls(vec![ToolCallItem {
-            name: "my-tool".into(),
-            params: serde_json::json!({}),
-            id: Some("call_1".into()),
-        }]),
-        ChatHistoryMessage::ToolResult {
-            name: "my-tool".into(),
-            content: "result".into(),
-        },
-        // Spurious extra result with no matching call
-        ChatHistoryMessage::ToolResult {
-            name: "skill-learner".into(),
-            content: "injected".into(),
-        },
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            content: "done".into(),
-        },
-    ];
-    crate::history::sanitize_history(&mut history);
-    // The extra ToolResult should be dropped
-    assert_eq!(history.len(), 3);
-    assert!(matches!(
-        &history[0],
-        ChatHistoryMessage::AssistantToolCalls(_)
-    ));
-    assert!(matches!(
-        &history[1],
-        ChatHistoryMessage::ToolResult { name, .. } if name == "my-tool"
-    ));
-    assert!(matches!(
-        &history[2],
-        ChatHistoryMessage::Text {
-            role: ChatRole::Assistant,
-            ..
-        }
-    ));
-}
-
-// ── value_to_params_map tests ─────────────────────────────────────────────────
-
-#[test]
-fn value_to_params_map_converts_object() {
-    let val = json!({"foo": "bar", "n": 42});
-    let map = super::value_to_params_map(&val);
-    assert_eq!(
-        map.len(),
-        2,
-        "object with two keys should produce two params"
-    );
-    assert_eq!(map["foo"], json!("bar"), "foo should map to \"bar\"");
-    assert_eq!(map["n"], json!(42), "n should map to 42");
-}
-
-#[test]
-fn value_to_params_map_empty_object() {
-    let val = json!({});
-    let map = super::value_to_params_map(&val);
-    assert!(map.is_empty(), "empty object should produce empty map");
-}
-
-#[test]
-fn value_to_params_map_non_object_returns_empty() {
-    for val in [json!(null), json!("string"), json!(42), json!([1, 2])] {
-        let map = super::value_to_params_map(&val);
-        assert!(map.is_empty(), "non-object {val} should produce empty map");
-    }
-}
-
-// ── Thinking persistence tests ────────────────────────────────────────────────
-
-/// Ollama response with empty content but non-empty thinking — triggers
-/// LlmResponse::Thinking in the LLM client parser.
-fn ollama_thinking(thought: &str) -> Value {
-    json!({
-        "model": "test",
-        "message": { "role": "assistant", "content": "", "thinking": thought },
-        "done": true
-    })
-}
-
-// ── Max-iterations error tests ────────────────────────────────────────────────
-
-#[tokio::test]
-async fn run_turn_max_iterations_returns_error() {
-    let server = MockServer::start().await;
-
-    // LLM always returns tool calls, never a final answer.
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(ollama_tool_calls(&["unknown-tool"])),
-        )
-        .mount(&server)
-        .await;
-
-    let mut config = AssistantConfig::default();
-    config.memory.enabled = false;
-    config.llm.max_iterations = 2;
-    let (orch, _) = build_with_config(&server.uri(), config).await;
-
-    let result = orch
-        .run_turn(
-            "trigger loop",
-            Uuid::new_v4(),
-            Interface::Cli,
-            None,
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await;
-
-    match result {
-        Ok(_) => panic!("should fail when max iterations reached"),
-        Err(e) => {
-            let err_msg = e.to_string();
-            assert!(
-                err_msg.contains("Max iterations"),
-                "error should mention max iterations: {err_msg}"
-            );
-        }
-    }
-}
-
-// ── Identity threading test ──────────────────────────────────────────────
-
-/// A mock tool that captures the [`ExecutionContext`] it receives.
-struct IdentityCaptureTool {
-    captured: Arc<tokio::sync::Mutex<Option<ExecutionContext>>>,
-}
-
-#[async_trait::async_trait]
-impl ToolHandler for IdentityCaptureTool {
-    fn name(&self) -> &str {
-        "identity-probe"
-    }
-    fn description(&self) -> &str {
-        "Captures execution context for testing"
-    }
-    fn params_schema(&self) -> Value {
-        json!({ "type": "object", "properties": {}, "required": [] })
-    }
-    async fn run(
-        &self,
-        _params: HashMap<String, Value>,
-        ctx: &ExecutionContext,
-    ) -> anyhow::Result<ToolOutput> {
-        *self.captured.lock().await = Some(ctx.clone());
-        Ok(ToolOutput::success("captured"))
-    }
-}
-
-#[tokio::test]
-async fn turn_identity_reaches_tool_handler() {
-    let server = MockServer::start().await;
-
-    // First LLM call → invoke our probe tool; second → final answer.
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(ollama_tool_calls(&["identity-probe"])),
-        )
-        .up_to_n_times(1)
-        .mount(&server)
-        .await;
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(ollama_answer("done")))
-        .mount(&server)
-        .await;
-
-    let (orch, _storage) = build(&server.uri()).await;
-
-    // Register our probe tool.
-    let captured = Arc::new(tokio::sync::Mutex::new(None));
-    let probe = Arc::new(IdentityCaptureTool {
-        captured: captured.clone(),
-    });
-    orch.executor.register_ambient_tool(probe);
-
-    // Run a turn with explicit identity.
-    let identity = TurnIdentity {
-        user_id: Some(UserId::from("usr_alice")),
-        org_id: Some(OrgId::from("org_acme")),
-        space_id: Some(SpaceId::from("spc_default")),
-    };
-
-    let result = orch
-        .run_turn(
-            "probe identity",
-            Uuid::new_v4(),
-            Interface::Cli,
-            None,
-            vec![],
-            identity,
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(result.answer, "done");
-
-    // Verify the tool handler received the correct identity.
-    let ctx = captured.lock().await;
-    let ctx = ctx.as_ref().expect("tool should have been called");
-    assert_eq!(
-        ctx.user_id.as_ref().map(|u| u.0.as_str()),
-        Some("usr_alice"),
-        "user_id should reach tool handler"
-    );
-    assert_eq!(
-        ctx.org_id.as_ref().map(|o| o.0.as_str()),
-        Some("org_acme"),
-        "org_id should reach tool handler"
-    );
-    assert_eq!(
-        ctx.space_id.as_ref().map(|s| s.0.as_str()),
-        Some("spc_default"),
-        "space_id should reach tool handler"
-    );
-}
-
-// ── Slice C: turn_had_errors signal ───────────────────────────────────────
-
-/// Tool that always returns an `Err(...)` — used to drive the
-/// `turn_had_errors` signal in the orchestrator turn loop.
-struct FailingTool;
-
-#[async_trait]
-impl ToolHandler for FailingTool {
-    fn name(&self) -> &str {
-        "failing-tool"
-    }
-
-    fn description(&self) -> &str {
-        "always errors"
-    }
-
-    fn params_schema(&self) -> Value {
-        json!({ "type": "object", "properties": {}, "required": [] })
-    }
-
-    async fn run(
-        &self,
-        _params: HashMap<String, Value>,
-        _ctx: &ExecutionContext,
-    ) -> anyhow::Result<ToolOutput> {
-        anyhow::bail!("simulated tool failure")
-    }
-}
-
-#[tokio::test]
-async fn run_turn_marks_had_errors_when_tool_fails() {
-    let server = MockServer::start().await;
-
-    // 1st LLM call: model invokes the failing tool.
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(ollama_tool_calls(&["failing-tool"])),
-        )
-        .up_to_n_times(1)
-        .mount(&server)
-        .await;
-
-    // 2nd LLM call: final answer (model recovers from the tool error).
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(ollama_answer("recovered")))
-        .mount(&server)
-        .await;
-
-    let (orch, _, executor) = build_with_executor(&server.uri()).await;
-    executor.register_ambient_tool(Arc::new(FailingTool));
-
-    let result = orch
-        .run_turn(
-            "trigger error",
-            Uuid::new_v4(),
-            Interface::Cli,
-            None,
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(
-        result.answer, "recovered",
-        "turn should still return the recovered final answer after a failing tool"
-    );
-    assert!(
-        result.had_errors,
-        "TurnResult.had_errors should be true after a tool returned Err"
-    );
-}
-
-#[tokio::test]
-async fn run_turn_clears_had_errors_when_no_tool_fails() {
-    let server = MockServer::start().await;
-    mount_answer(&server, "all good").await;
-
-    let (orch, _, _executor) = build_with_executor(&server.uri()).await;
-
-    let result = orch
-        .run_turn(
-            "no tool calls",
-            Uuid::new_v4(),
-            Interface::Cli,
-            None,
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await
-        .unwrap();
-
-    assert_eq!(
-        result.answer, "all good",
-        "turn without tool failures should return the model answer"
-    );
-    assert!(
-        !result.had_errors,
-        "TurnResult.had_errors should be false when no tool errored"
-    );
-}
-
-#[tokio::test]
-async fn run_turn_with_tools_marks_had_errors_when_extension_fails() {
-    // Exercises the Slack/Mattermost extension-tool path: a failing extension
-    // handler must propagate `had_errors = true` through
-    // `handle_final_answer_with_extensions` and the per-turn accumulator,
-    // mirroring `run_turn_marks_had_errors_when_tool_fails` for `run_turn`.
-    let server = MockServer::start().await;
-
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(
-            ResponseTemplate::new(200).set_body_json(ollama_tool_calls(&["failing-tool"])),
-        )
-        .up_to_n_times(1)
-        .mount(&server)
-        .await;
-
-    Mock::given(method("POST"))
-        .and(path("/api/chat"))
-        .respond_with(ResponseTemplate::new(200).set_body_json(ollama_answer("recovered")))
-        .mount(&server)
-        .await;
-
-    let (orch, _) = build(&server.uri()).await;
-    let failing_ext: Arc<dyn ToolHandler> = Arc::new(FailingTool);
-
-    // Use Cli interface so a plain final-answer turn is accepted without a
-    // `reply` extension tool; the focus of this test is the extension-handler
-    // failure path, not the messenger reply protocol.
-    //
-    // `run_turn_with_tools_impl` always returns `answer: String::new()` (the
-    // assistant text is persisted to the conversation store, not surfaced
-    // through TurnResult), so we only assert on `had_errors`.
-    let result = orch
-        .run_turn_with_tools(
-            "trigger error",
-            Uuid::new_v4(),
-            Interface::Cli,
-            vec![failing_ext],
-            None,
-            vec![],
-            vec![],
-            TurnIdentity::default(),
-        )
-        .await
-        .unwrap();
-
-    assert!(
-        result.had_errors,
-        "TurnResult.had_errors must be true after a failing extension tool \
-         (covers the run_turn_with_tools_impl wiring)"
     );
 }
